fixed projecthighlight export
